moving all the java classes to a java package
