+++ src/main/java/com/novoda/notils/java/caster/Collections.java
@@ -1,4 +1,4 @@
-package com.novoda.notils.caster;
+package com.novoda.notils.java.caster;
 
 import java.util.*;
