+++ src/main/java/com/novoda/notils/java/string/StringUtils.java
@@ -1,4 +1,4 @@
-package com.novoda.notils.string;
+package com.novoda.notils.java.string;
 
 import java.util.Collection;
 import java.util.Iterator;
